working setup

+++ UsbSerial/lib/logHandler/LogHandler.h
@@ -6,7 +6,9 @@
 
 class LogHandler {
 public:
-    LogHandler(LogHandlerCommunication &communication, PressureSensor &frontPressureSensor, PressureSensor &backPressureSensor, PressureSensor &tankPressureSensor);
+    LogHandler(LogHandlerCommunication &communication, PressureSensor &frontPressureSensor,
+               PressureSensor &backPressureSensor, PressureSensor &tankPressureSensor);
+
     ~LogHandler();
 
     void StartFrontLog(bool togetherMove = false);
@@ -20,16 +22,18 @@ public:
     void SendLog();
 
 private:
-    String CreateLogMessage(String message, double startPressure, double endPressure, double startTankPressure, long time, bool direction, bool togetherMove);
+    String CreateLogMessage(String message, float startPressure, float endPressure, float startTankPressure, long time,
+                            bool direction, bool togetherMove);
+
     unsigned long startTimeFront = 0;
-    double startPressureFront = 0;
-    double startTankPressureFront = 0;
+    float startPressureFront = 0;
+    float startTankPressureFront = 0;
     bool togetherMoveFront = false;
     bool sendLogFront = false;
 
     unsigned long startTimeBack = 0;
-    double startPressureBack = 0;
-    double startTankPressureBack = 0;
+    float startPressureBack = 0;
+    float startTankPressureBack = 0;
     bool togetherMoveBack = false;
     bool sendLogBack = false;
